% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{codeDemo}
\alias{codeDemo}
\title{Code demonstration.}
\usage{
codeDemo()
}
\description{
Run through a quick demonstration of the SLGACloud functions for accessing the SLGA COGs Datastore URLs
}
\author{
Ross Searle
}
