% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{codeDemoCOGs}
\alias{codeDemoCOGs}
\title{COGS code demonstration.}
\usage{
codeDemoCOGs()
}
\description{
Run through a quick demonstration of using COGS in R with the "terra" package.
}
\author{
Ross Searle
}
