% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{cogDownload}
\alias{cogDownload}
\title{Download a COG file}
\usage{
cogDownload(url = NULL, dest = "", quiet = FALSE)
}
\arguments{
\item{url}{Web location of the Cloud Optimised Geotiff}

\item{dest}{Download file location}

\item{quiet}{Show file download progress bar if TRUE}
}
\value{
Cloud Optimised Geotiff
}
\description{
Download an entire COG Geotiff file to your local machine
}
\examples{
cogDownload(url = 'https://esoil.io/TERNLandscapes/Public/Products/TERN/Covariates/Mosaics/90m/Veg_Landsat8Bare2.tif', dest='c:/temp/demoCOG.tif')
}
\author{
Ross Searle
}
