% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{cogLoad}
\alias{cogLoad}
\title{Load a raster dataset from a COG URL}
\usage{
cogLoad(url = NULL)
}
\arguments{
\item{url}{Web location of the Cloud Optimised Geotiff}
}
\value{
SpatRaster
}
\description{
Uses the 'terra::rast' function and prepends the '/vsicurl/' bit to save you doing it manually
}
\examples{
\dontrun{
cogLoad('https://esoil.io/TERNLandscapes/Public/Products/TERN/Covariates/Mosaics/90m/Veg_Landsat8Bare2.tif')
}
}
\author{
Ross Searle
}
