% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{cogPlot}
\alias{cogPlot}
\title{Plot COG files}
\usage{
cogPlot(url)
}
\arguments{
\item{url}{URL of the Cloud Optimised Geotiff.}
}
\value{
plot
}
\description{
Plot a single Cloud Optimised Geotiff.
}
\details{
A method for quickly viewing a COG.
}
\examples{
cogPlot(url)
}
\author{
Ross Searle
}
