% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{cogPreview}
\alias{cogPreview}
\title{Preview a list of COG files}
\usage{
cogPreview(urls = NULL, delay = 3)
}
\arguments{
\item{urls}{List of Cloud Optimised Geotiff URLs}

\item{delay}{Delay in seconds between drawing plots. Minimum = 3 seconds}
}
\value{
plot
}
\description{
Iterate through a list of Cloud Optimised Geotiff URLs and plot them.
}
\details{
A method for quickly viewing data products.
}
\examples{
\dontrun{
cogs <- getProductMetaData(Detail = 'Low',  Attribute='Parent_Material', Resolution = '90m')
cogPreview(urls = cogs$COGsPath[1:6],  delay = 5)
}
}
\author{
Ross Searle
}
