% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStore.R
\name{COGSDataStoreURLs}
\alias{COGSDataStoreURLs}
\title{URLs for the TERN Landscapes COGs Data Store}
\usage{
COGSDataStoreURLs()
}
\value{
dataframe
}
\description{
List relevant URLs for the TERN Landscapes COGs Data Store
}
