% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariates.R
\name{covariateCOGsURLs}
\alias{covariateCOGsURLs}
\title{Returns a dataframe of available covariate rasters and their COGs URLs}
\usage{
covariateCOGsURLs(product = NULL, layername = NULL)
}
\arguments{
\item{product}{The product to return names for. Options are '90mCovariates' & '30mCovariates'}

\item{layername}{Filter on a specific layer name.}
}
\value{
dataframe
}
\description{
Returns a dataframe of available covariate rasters and their COGs URLs
}
\examples{
covariateCOGsURLs(product = '90mCovariates', layername = 'Veg_Landsat8Bare2')
covariateCOGsURLs(product = '30mCovariates', layername = NULL)
}
