% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariates.R
\name{covariateNames}
\alias{covariateNames}
\title{List of covariate raster names}
\usage{
covariateNames(product = "90mCovariates")
}
\arguments{
\item{product}{The product to return names for. Options are '90mCovariates' & '30mCovariates'}
}
\value{
list
}
\description{
Returns a list of available covariate raster names
}
\examples{
covariateNames(product = '90mCovariates')
}
