% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{downloadcog}
\alias{downloadcog}
\title{Download a COG file}
\usage{
downloadcog(url = NULL, dest = "")
}
\arguments{
\item{url}{Web location of the Cloud Optimised Geotiff}

\item{dest}{Download file location}
}
\value{
file
}
\description{
Download an entire COG Geotiff file to your local machine
}
\examples{
downloadcog(url = 'https://esoil.io/TERNLandscapes/Public/Products/TERN/Covariates/Mosaics/90/Veg_Landsat8Bare2.tif', dest='c:/temp/demoCOG.tif')
}
