% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{drillRaster}
\alias{drillRaster}
\title{Extract values from COGS files at a location}
\usage{
drillRaster(COGPath, Longitude, Latitude)
}
\arguments{
\item{COGPath}{The web path to the Cloud Optimised GeoTIFF}

\item{Longitude}{Longitude value - needs to be with Australia}

\item{Latitude}{Latitude value - needs to be with Australia}
}
\value{
data.frame
}
\description{
Drill SLGA rasters at a location.
}
\details{
Drill the Cloud Optimised GeoTIFF rasters in the SLGA data store at a given location and return values for that location.
You need to use the "StagingPath" returned from the getProductMetaData() function as the "Path" variable in this function
}
\examples{
\dontrun{
rasters <- getProductMetaData(Detail = 'High', Attribute = 'Sand', Component = 'Modelled-Value',  isCurrentVersion = 1 )

for(i in 1:nrow(l)){
  v <- drillRaster(rasters$StagingPath[i], Longitude, Latitude)
  print(v)
}
}
}
\author{
Ross Searle
}
