% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterMethods.R
\name{drillRasters}
\alias{drillRasters}
\title{Extract values from COGS files at a location}
\usage{
drillRasters(Products, Longitude, Latitude, Verbose = T)
}
\arguments{
\item{Products}{Dataframe of product records returned from the getProductMetaData() function}

\item{Longitude}{Longitude value - needs to be with Australia}

\item{Latitude}{Latitude value - needs to be with Australia}

\item{Verbose}{return all metadata or just the name and value}
}
\value{
data.frame
}
\description{
Drill SLGA rasters at a location.
}
\details{
Drill the Cloud Optimised GeoTIFF rasters in the SLGA data store at a given location and return values for that location.
}
\examples{
 prods <- getProductMetaData(Detail = 'Low',  Attribute='Parent_Material', Resolution = '90m')
drillRasters(Products = prods[1:3,], Longitude = 151, Latitude = -26, Verbose = F)
}
\author{
Ross Searle
}
