% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea_rasSp_fast.R
\name{ea_rasSp_fast}
\alias{ea_rasSp_fast}
\title{Fits a mass-preserving spline model to a multi-layer soil property
\code{SpatRaster} and returns raster layers containing spline-averaged values
over user-defined depth intervals. This function uses precomputed spline matrices
for efficiency and applies the spline cell-wise using \code{terra::app()}.}
\usage{
ea_rasSp_fast(
  obj,
  lam = 0.1,
  dIn = c(0, 5, 15, 30, 60, 100, 200),
  dOut = c(0, 30, 60),
  vlow = 0,
  vhigh = 100,
  depth_res = 1
)
}
\arguments{
\item{obj}{Object of class \code{SpatRaster} with one layer per input depth interval.}

\item{lam}{Spline stiffness parameter (\eqn{\lambda}); smaller values = smoother splines.}

\item{dIn}{Numeric vector of input depth boundaries (e.g. \code{c(0,5,15,...)}).}

\item{dOut}{Numeric vector defining output depth intervals (e.g. \code{c(0,30,60)}).}

\item{vlow}{Minimum bound to truncate spline predictions (default = 0).}

\item{vhigh}{Maximum bound to truncate spline predictions (default = 100).}

\item{depth_res}{Numeric; depth interpolation resolution (e.g. 1 = 1cm, 5 = 5cm).}
}
\value{
A \code{SpatRaster} with one layer per output depth interval.
}
\description{
Fits a mass-preserving spline model to a multi-layer soil property
\code{SpatRaster} and returns raster layers containing spline-averaged values
over user-defined depth intervals. This function uses precomputed spline matrices
for efficiency and applies the spline cell-wise using \code{terra::app()}.
}
\examples{
\dontrun{
library(terra)

clay_urls <- c(
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_000_005_EV_N_P_AU_TRN_N_20210902.tif',
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_005_015_EV_N_P_AU_TRN_N_20210902.tif',
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_015_030_EV_N_P_AU_TRN_N_20210902.tif',
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_030_060_EV_N_P_AU_TRN_N_20210902.tif',
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_060_100_EV_N_P_AU_TRN_N_20210902.tif',
  '/vsicurl/https://esoil.io/TERNLandscapes/Public/Products/TERN/SLGA/CLY/CLY_100_200_EV_N_P_AU_TRN_N_20210902.tif'
)

clay_stack <- terra::rast(clay_urls)
aoi <- ext(149.00, 149.10, -36.00, -35.90)
clay_crop <- terra::crop(clay_stack, aoi)

out <- ea_rasSp_fast(
  obj = clay_crop,
  lam = 0.1,
  dIn = c(0, 5, 15, 30, 60, 100, 200),
  dOut = c(0, 30, 60),
  depth_res = 2
)

plot(out)
}

}
\author{
Brendan Malone
}
\keyword{methods}
