% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getParameterValues}
\alias{getParameterValues}
\title{Get allowable query parameter values}
\usage{
getParameterValues(Parameter = "Name")
}
\arguments{
\item{Parameter}{The allowable parameters that can be filtered on are 'Attribute', 'DataType', 'Product', 'Source', 'Component', 'Name' and  'Resolution'}
}
\value{
data.frame
}
\description{
Returns a tibble of allowable query parameter values
}
\details{
This function provides the allowable values for filtering metadata records on.
}
\examples{
getParameterValues(Parameter = 'Product')
getParameterValues(Parameter = 'Source')
}
\author{
Ross Searle
}
