% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{getProductMetaData}
\alias{getProductMetaData}
\title{Get Product MetaData}
\usage{
getProductMetaData(
  Detail = "High",
  Product = NULL,
  DataType = NULL,
  Source = NULL,
  Attribute = NULL,
  Component = NULL,
  Name = NULL,
  Resolution = NULL,
  Version = NULL,
  isCurrentVersion = NULL
)
}
\arguments{
\item{Detail}{The amount of information to return. Options are 'Low' (Name &  COGsPath),'Moderate', (Name, COGsPath, UpperDepth_m, LowerDepth_m & Units) and High (All metadata)}

\item{Product}{Filter by Product ('SLGA', '30m_Covariate' or '90m_Covariate').}

\item{DataType}{Filter by DataType ('Soil' or 'DSM_Covariate').}

\item{Source}{Filter by Source - who generated the data.}

\item{Attribute}{Filter by Attribute.}

\item{Component}{Filter by Component (' Modelled-Value', 'Lower-CI', 'Upper-CI', 'Value').}

\item{Name}{Filter by the data set name.}

\item{Resolution}{Filter by spatial resolution ('90m' or '30m').}

\item{Version}{Filter by version number (current options are 1 or 2).}

\item{isCurrentVersion}{return only the latest version of a product (1 for True and 0 for False).}
}
\value{
data.frame
}
\description{
Returns a tibble of available SLGA product metadata
}
\details{
The datastore can be queried to return filtered lists of metadata. The allowable values for filtering can be accessed using the getParameterValues() function
}
\examples{
getProductMetaData(Detail = 'High', Product='SLGA', Attribute='Available Water Capacity')
}
\author{
Ross Searle
}
