% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COGsUtils.R
\name{loadcog}
\alias{loadcog}
\title{Load a raster dataset from a COG URL}
\usage{
loadcog(url = NULL)
}
\arguments{
\item{url}{Web location of the Cloud Optimised Geotiff}
}
\value{
SpatRaster
}
\description{
Load a raster dataset from a COG URL
}
\examples{
loadcog('https://esoil.io/TERNLandscapes/Public/Products/TERN/Covariates/Mosaics/90/Veg_Landsat8Bare2.tif')
}
