% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websites.R
\name{openCogsWebsites}
\alias{openCogsWebsites}
\title{Open the various web pages in the TERN Landscapes COGs Data Store}
\usage{
openCogsWebsites(product = "COGsMainPage", type = "Formatted")
}
\arguments{
\item{product}{Which website to open. Options are 'SLGA', 'COGsMainPage', '90mCovariates', '30mCovariates'}

\item{type}{Go to the nicely formatted web page or the raw http file listing. Options are 'Formatted' & 'Raw'}
}
\description{
Open the various web pages in the TERN Landscapes COGs Data Store
}
\examples{
openCogsWebsites('COGsMainPage')
openCogsWebsites('90mCovariates', type='Formatted')
openCogsWebsites('30mCovariates', type='Raw')
openCogsWebsites('SLGA')
}
