% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStore.R
\name{pingDataStore}
\alias{pingDataStore}
\title{Check that the SLGA Data Store is available}
\usage{
pingDataStore()
}
\value{
string
}
\description{
This function hits the SLGA Data Store to check that it is available
}
