% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLGACOGs.R
\name{slgaNames}
\alias{slgaNames}
\title{List of available SLGA raster names}
\usage{
slgaNames(product = "CLY")
}
\arguments{
\item{product}{The product to return names for. Options are "AWC", "BDW", "CLY", "DER", "DES", "ECE", "NTO", "pHc", "PTO", "SLT", "SND", "SOC")}
}
\value{
list
}
\description{
Returns a list of available SLGA raster names
}
\examples{
SLGANames(product='CLY')
}
