% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websites.R
\name{SLGAWebsites}
\alias{SLGAWebsites}
\title{SLGAs Websites}
\usage{
SLGAWebsites(WebSiteName = NULL)
}
\arguments{
\item{WebSiteName}{If a name is supplied the URL will be opened in your browser.}
}
\value{
dataframe
}
\description{
Locations of the various web sites relevant to the Soil and Landscape Grid of Australia.
}
\details{
If no WebSiteName value is supplied, a list of urls will be returned. If a  WebSiteName parameter value is supplied the website will be opened in your browser
}
\examples{
SLGAWebsites()
SLGAWebsites('TERN Landscapes COGs Datastore Main Page')
}
\author{
Ross Searle
}
