% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covariates.R
\name{covariateMetaData}
\alias{covariateMetaData}
\title{Return the metadata information for the raster covariate stack}
\usage{
covariateMetaData(product = NULL)
}
\arguments{
\item{product}{Filter on the covariate stack for a particular product stack. Options are 30mCovariates or 90mCovariates}
}
\value{
dataframe
}
\description{
Return the metadata information for the raster covariate stack
}
\examples{
covariateMetaData()
covariateMetaData(product = '90mCovariates')
}
