% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLGACOGs.R
\name{slgaCodes}
\alias{slgaCodes}
\title{List available SLGA product codes}
\usage{
slgaCodes()
}
\value{
dataframe
}
\description{
Returns a list of available SLGA product codes with it plain English description
}
\examples{
SLGACodes()
}
